srcs/backend/requirements.txt
Django==5.1.3
django-axes==7.0.1
djangorestframework==3.15.2
asgiref==3.8.1
pillow==11.0.0
sqlparse==0.5.2
typing_extensions==4.12.2